Restructure project according to choo-cli
