+++ lib/blockly/index.js
