+++ lib/utils/assets.js
