+++ lib/utils/clock.js
