+++ lib/utils/game.js
