+++ lib/utils/lodash-extension.js
