+++ lib/utils/pathfinder.js
@@ -1,5 +1,5 @@
 const { AStarFinder, Grid } = require('pathfinding')
-const initalState = require('../game/initial-state')
+const initialState = require('../../models/game/initial-state')
 const { ORIENTATION } = require('./types')
 
 const finder = new AStarFinder()
@@ -23,7 +23,8 @@ function getGrid (state) {
 }
 
 function getPath (startPos, endPos) {
-  const grid = getGrid(initalState)
+  // TODO: remove this reference to initial state, instead pass game state in to calculate grid
+  const grid = getGrid(initialState)
   return finder.findPath(startPos.x, startPos.y, endPos.x, endPos.y, grid)
 }
