+++ lib/utils/robot-runtime.js
@@ -1,7 +1,7 @@
 const _ = require('lodash')
 const uid = require('uid')
 const esper = require('esper.js')
-const { getGameState, getEntity } = require('../utils/game')
+const { getGameState, getEntity } = require('./game')
 
 class RobotRuntime {
