+++ models/game/components/collector.js
