+++ models/game/components/spawner.js
